+++ lib/gds_api/local_links_manager.rb
@@ -1,9 +1,8 @@
 require_relative 'base'
 
 class GdsApi::LocalLinksManager < GdsApi::Base
-  def local_link(authority_slug, lgsl, lgil = nil)
-    url = "#{endpoint}/api/link?authority_slug=#{authority_slug}&lgsl=#{lgsl}"
-    url += "&lgil=#{lgil}" if lgil
+  def local_link(authority_slug, lgsl, lgil)
+    url = "#{endpoint}/api/link?authority_slug=#{authority_slug}&lgsl=#{lgsl}&lgil=#{lgil}"
     get_json(url)
   end